фикстуры помещал на нижние уровни
